Add NavBar component to set precedent and suivant button on it, and it's fonctionnal!
